requirements/prod.txt
-r common.txt
Flask-SSLify==0.1.4
gunicorn==18.0